+++ src/jdom/jacdac-jdom.ts
@@ -14,10 +14,6 @@ export * from "./register"
 export * from "./field"
 export * from "./event"
 export * from "./bus"
-export * from "./proto"
-export * from "./workerproto"
-export * from "./hf2"
-export * from "./usb"
 export * from "./logparser"
 export * from "./pretty"
 export * from "./flashing"
@@ -46,8 +42,14 @@ export * from "./serviceprovider"
 export * from "./controlserver"
 export * from "./color"
 export * from "./makecode"
-export * from "./transport"
-export * from "./bluetooth"
 export * from "./rolemanagerclient"
 export * from "./qualityofservice"
 export * from "./bridge"
+export * from "./transport/transport"
+export * from "./transport/bluetooth"
+export * from "./transport/usbio"
+export * from "./transport/proto"
+export * from "./transport/hf2"
+export * from "./transport/microbit"
+export * from "./transport/workertransport"
+export * from "./transport/usb"
