+++ src/jdom/transport/microbit.ts
@@ -1,5 +1,5 @@
-import { Transport } from "./hf2"
 import Proto from "./proto"
+import USBIO from "./usbio"
 import {
     delay,
     PromiseQueue,
@@ -10,7 +10,7 @@ import {
     fromHex,
     bufferConcat,
     fromUTF8,
-} from "./utils"
+} from "../utils"
 
 interface SendItem {
     buf: Uint8Array
@@ -30,7 +30,7 @@ export class CMSISProto implements Proto {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     private _lastInterval: any
 
-    constructor(private io: Transport) {
+    constructor(private io: USBIO) {
         console.log(`micro:bit: start proto`)
     }
