+++ src/serviceworker/transportproxy.ts
@@ -1,0 +1,5 @@
+export default interface TransportProxy {
+    connect(): Promise<void>
+    send(payload: Uint8Array): Promise<void>
+    disconnect(): Promise<void>
+}
